requirements.txt
# ====================================
# StoryBook 项目依赖
# ====================================

# LangGraph 核心（必需）
langgraph==0.2.55
langchain==0.3.13
langchain-core==0.3.26

# HTTP 请求（必需）
requests>=2.32.0            # 用于调用 API

# 配置管理（必需）
python-dotenv==1.0.0        # 环境变量管理
pydantic==2.10.4            # 数据验证
typing-extensions==4.12.2   # 类型提示扩展

# 图像处理（可选，用于生成占位图）
Pillow>=10.0.0              # 图像处理

# ====================================
# 注意：不再需要厂商 SDK！
# ====================================
#
# ❌ 已移除：anthropic, openai, dashscope
# ✅ 改用：直接 HTTP 请求
#
# 优势：
# - 只需配置 URL + Key + Model
# - 无需安装特定 SDK
# - 支持任何兼容的服务
# - 更灵活、更通用
#
# ====================================

# 最小安装（只安装必需的）：
#   pip install langgraph langchain langchain-core requests python-dotenv pydantic

# 完整安装（包括可选的）：
#   pip install -r requirements.txt
